Support linker args when building in a static project.
This change looks for the `LIBARCHIVE_LDFLAGS` environment variable, but
only when `LIBARCHIVE_STATIC` is set. This allows us to pass the libaray
path for some of libarchive's dependencies (assuming they are built in),
such as zlib, bzip2, xz, openssl, etc. The `-L` entries are a path
containing the static archive (i.e. `*.a`) and the `-l` entries are the
specific libraries to link against.

Here's an example invocation of `cargo build`:

```sh
env \
  LIBARCHIVE_LIB_DIR=/usr/local/musl/libarchive/lib \
  LIBARCHIVE_INCLUDE_DIR=/usr/local/musl/libarchive/include \
  LIBARCHIVE_LDFLAGS="-L/usr/local/musl/xz/lib -llzma -L/usr/local/musl/zlib/lib -lz" \
  LIBARCHIVE_STATIC=true \
  cargo build --target=x86_64-unknown-linux-musl
```

Without knowing where these libararies live, the linker only finds
missing symbols when assembing the final binary. Don't even ask how I've
figure this out... ;)